Move feature flag initialization logic out of lib.rs.
